refactor: split panels.py into panels/ package
Split the monolithic panels.py module into a package with one file
per panel component, following the same pattern used for models/ and
services/:

- panels/stock_summary.py — StockSummaryPanel
- panels/low_stock.py — LowStockPanel
- panels/recent_movements.py — RecentMovementsPanel
- panels/__init__.py — re-exports all panel classes

Updated ARCHITECTURE.md project structure tree to reflect the new
panels/ package layout and added filters.py to the listing.
