+++ inventory/panels/low_stock.py
@@ -1,0 +1,33 @@
+"""Low-stock items dashboard panel."""
+
+from django.db.models import F
+from wagtail.admin.ui.components import Component
+
+from inventory.models import StockRecord
+
+
+class LowStockPanel(Component):
+    """Dashboard table of products below their reorder point.
+
+    Shows up to 10 items sorted by deficit (most critical first),
+    with a link to the full low-stock alerts view.
+    """
+
+    name = "low_stock"
+    template_name = "inventory/panels/low_stock.html"
+    order = 110
+
+    def get_context_data(self, parent_context=None):
+        context = super().get_context_data(parent_context)
+        low_stock_records = (
+            StockRecord.objects
+            .filter(
+                product__reorder_point__gt=0,
+                quantity__lte=F("product__reorder_point"),
+            )
+            .select_related("product", "location")
+            .annotate(deficit=F("product__reorder_point") - F("quantity"))
+            .order_by("-deficit")[:10]
+        )
+        context["low_stock_records"] = low_stock_records
+        return context
